fix(circleprogress): demo拆解与规范
