+++ src/packages/circleprogress/demos/h5/demo1.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { CircleProgress } from '@nutui/nutui-react'
+
+const Demo1 = () => {
+  return (
+    <>
+      <CircleProgress percent={20} />
+      <CircleProgress percent={60}>60%</CircleProgress>
+    </>
+  )
+}
+export default Demo1
