+++ src/packages/circleprogress/demos/h5/demo2.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { CircleProgress } from '@nutui/nutui-react'
+
+const Demo2 = () => {
+  return (
+    <>
+      <CircleProgress percent={50} strokeWidth={2} />
+      <CircleProgress percent={60} strokeWidth={10} background="#e5e9f2" />
+    </>
+  )
+}
+export default Demo2
