+++ src/packages/circleprogress/demos/h5/demo4.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { CircleProgress } from '@nutui/nutui-react'
+
+const Demo4 = () => {
+  return (
+    <>
+      <CircleProgress percent={50} radius={60}>
+        50%
+      </CircleProgress>
+    </>
+  )
+}
+export default Demo4
