+++ src/packages/circleprogress/demos/taro/demo2.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { CircleProgress } from '@nutui/nutui-react-taro'
+
+const Demo2 = () => {
+  return (
+    <>
+      <CircleProgress percent={50} strokeWidth={10} />
+    </>
+  )
+}
+export default Demo2
